+++ tutoGitSFML/tutoGitSFML/tutoGitSFML.vcxproj.filters
@@ -26,5 +26,8 @@
     <ClInclude Include="..\..\headers\saisieSecurisee.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\headers\pieces.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
